Add files via upload
This project is about demonstrating the most efficient way for a flower to distribute its seeds. By looking at a model of seeds represented as points each turning some fraction of 2π, we get really amazing shapes. When the get to a certain number, called the golden ratio, we can look at how efficiently those seeds are being spread across the flower, and quite frighteningly, the distribution matches with the head of a sunflower in the real world.
Inspiration: https://www.youtube.com/watch?v=sj8Sg8qnjOg